Db 3 0 (#65)
DB 3.0 Migration including a bunch of qol changes.

Add categories
Add UUID support for lists, entries, training progress
List UUID tombstones to allow online syncing
Entry UUID tombstones to allow online syncing
Category UUID tombstones to allow online syncing
Changed date for lists,entries and categories
Replace manual auto-increment with the row-id hack of sqlite link
Replace dual-key for entries with one row-id based entry key
Add a per entry training history, to allow syncing
Use cascading delete and foreign keys mostly everywhere (except for session meta data that is essentially a string:string hashmap)
Remove KEY_CREATED and KEY_CORRECT from entries as only deltas can be synchronized and splitting online/offline user code isn't sustainable.
Migrate a bunch of things to kotlin while doing so and use auto parcelize for VEntry and VList, also enforcing changed + created dates.
Replace the legacy "table" and "voc" names with "list" and "entry".
Generic settings table that can be synced
Shared-Flag for lists
Upgrade path for old DBs
optimize entry tombstone by deleting them on list deletion
enforce non-null for foreign refs
migrate a bunch of tests to the new junit setup, fixing CI

+++ app/src/main/java/vocabletrainer/heinecke/aron/vocabletrainer/lib/Comparator/GenEntryComparator.java
@@ -8,7 +8,7 @@ import vocabletrainer.heinecke.aron.vocabletrainer.lib.Storage.VEntry;
  * can't handle more than one head entry
  */
 public class GenEntryComparator extends GenericComparator<VEntry, String> {
-    private final int headID;
+    private final long headID;
 
     /**
      * Generic VEntry Comparator
@@ -17,7 +17,7 @@ public class GenEntryComparator extends GenericComparator<VEntry, String> {
      *                   passed array order defines the comparision priority
      * @param headID     ID of entry to set on top
      */
-    public GenEntryComparator(ValueRetriever[] retrievers, final int headID) {
+    public GenEntryComparator(ValueRetriever[] retrievers, final long headID) {
         super(retrievers);
         this.headID = headID;
     }
